Simple CRUD done.

+++ demo/src/main/java/com/example/demo/repository/user_repos.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entities.user_entity;
 
 @Repository
-public interface user_repos extends JpaRepository<user_entity, Integer> {
+public interface user_repos extends JpaRepository<user_entity, Long> {
 //	@Override
 //	@Query(value = "select user.iduser, user.password, user.name, user.gender, user.dateCreated, user.status "
 //			+ "from pbl4.user as user",
